backend/requirements.txt
# FastAPI & Server
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# SQLModel & Database
sqlmodel==0.0.31
sqlalchemy[asyncio]==2.0.34
psycopg[binary,pool]==3.2.10
asyncpg==0.31.0

# Database Migrations
alembic==1.13.2

# Environment & Configuration
python-dotenv==1.2.1
pydantic-settings==2.11.0

# Authentication & Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==5.0.0

# Validation & Utilities
email-validator==2.3.0
python-dateutil==2.9.0.post0

# Data Processing
openpyxl==3.1.5
pandas==3.0.1

# Development & Testing
pytest==8.4.2
pytest-asyncio==1.3.0
httpx==0.28.1
